# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Hacktiv8 Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/api/v1",
    "paths": {
        "/ping": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.PingSwag"
                        }
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "description": "get all todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Get All TODOS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetTodosSuccessSwag"
                        }
                    }
                }
            },
            "post": {
                "description": "create todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Create TODO",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.CreateTodoSuccessSwag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/views.CreateTodoFailureSwag"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "get todos by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Get TODO By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ToDo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetTodoSuccessSwag"
                        }
                    }
                }
            },
            "put": {
                "description": "update todos by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Update TODO By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ToDo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.UpdateTodoSuccessSwag"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete todos by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOS"
                ],
                "summary": "Delete TODO By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ToDo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.DeleteTodoSuccessSwag"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.PingSwag": {
            "type": "object",
            "properties": {
                "health": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "request.CreateTodo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "request.UpdateTodo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "views.CreateTodoFailureSwag": {
            "type": "object",
            "properties": {
                "additional_info": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "error": "Title is required"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "BAD Request"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "views.CreateTodoPayload": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "views.CreateTodoSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "CREATE TODO SUCCESS"
                },
                "payload": {
                    "$ref": "#/definitions/views.CreateTodoPayload"
                },
                "status": {
                    "type": "integer",
                    "example": 201
                }
            }
        },
        "views.DeleteTodoSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "DELETE TODO BY ID SUCCESS"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "views.GetTodoPayload": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "views.GetTodoSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "GET TODO BY ID SUCCESS"
                },
                "payload": {
                    "$ref": "#/definitions/views.GetTodoPayload"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "views.GetTodosSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "GET ALL TODO SUCCESS"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.GetTodoPayload"
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "views.UpdateTodoPayload": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Desc TODO"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Title TODO"
                }
            }
        },
        "views.UpdateTodoSuccessSwag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "UPDATE TODO BY ID SUCCESS"
                },
                "payload": {
                    "$ref": "#/definitions/views.UpdateTodoPayload"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}